feat(series): implement series code and improve backup
The series code is stored in each series file, and it is now how series-specific QSettings can be stored. Auto-backup has been rebranded to Auto-versioning, and the backup (previously manual backup) dialog has had date and time delimiters added to it, which are stored in non-series-specific QSettings. The code for the FileDialog directory storage has also been fixed, as it was not properly storing the last visited folder all the time.