fix(mouse): account for multiple buttons
If user presses both left and right at the same time, right is favored. If the user weas in the middle of a scribble trace, it gets canceled. If the user presses the middle button with any other button, there are two cases: if panzooming, keep the middle button and do nothing, otherwise, ignore the middle button and do nothing