fix(lists): ensure obj attrs not deleted when modifying trace attrs
Object attributes were cleared when trace attributes of multiple
objects were changed simultaneously. This is because the program is
written to migrate object attributes when the name of an object is
changed. However, this applied when multiple objects where renamed.
This has been fixed by ensuring that a None value in the series
function that modifies the object is not considered a different name.