fix(field): implement hack to reinstate auto-merging
The new trace_function decorator passes all traces selected in the field
into functions it decorates, which is problematic for auto-merging. This
also seems to be messing with other arguments. (In this case,
merg_attrs_only argument in mergeTraces is also set to all selected traces,
which is evaluated as truthy in conditionals.) This changed the default
behavior of auto-merging to merge only attributes. To get around this, we
will implement a temporary hack that we will likely need to revisit.